feat: migrate Label component to theme architecture
Implemented registry pattern for Label component:
- Created themes/default/components/Label.tsx (Standard shadcn)
- Created themes/glass-refraction/components/Label.tsx (Glassmorphism)
- Refactored components/ui/label.tsx to dynamically load correct version

Glass Label features:
- Enhanced contrast (text-black dark:text-white/90)
- Subtle drop shadow for readability
- Consistent typography
